pie chart works
