+++ lib/laddu_seva/log.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+import 'package:synchronized/synchronized.dart';
+
+class Log extends StatefulWidget {
+  const Log({super.key});
+
+  @override
+  State<Log> createState() => _LogState();
+}
+
+// hint: LogKey.currentState!.refresh();
+final GlobalKey<_LogState> LogKey = GlobalKey<_LogState>();
+
+class _LogState extends State<Log> {
+  final _lockInit = Lock();
+
+  Future<void> _futureInit() async {
+    await _lockInit.synchronized(() async {
+      await Future.delayed(const Duration(seconds: 2));
+    });
+  }
+
+  Future<void> refresh() async {
+    await _futureInit();
+    setState(() {});
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<void>(
+      future: _futureInit(),
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return const Center(
+            child: CircularProgressIndicator(),
+          );
+        } else if (snapshot.hasError) {
+          return Text('Error: ${snapshot.error}');
+        } else {
+          return const Placeholder();
+        }
+      },
+    );
+  }
+}
