+++ lib/laddu_seva/summary.dart
@@ -1,0 +1,136 @@
+import 'package:fl_chart/fl_chart.dart';
+import 'package:flutter/material.dart';
+import 'package:garuda/fb.dart';
+import 'package:garuda/laddu_seva/datatypes.dart';
+import 'package:synchronized/synchronized.dart';
+
+class Summary extends StatefulWidget {
+  const Summary({super.key});
+
+  @override
+  State<Summary> createState() => _SummaryState();
+}
+
+// hint: templateKey.currentState!.refresh();
+final GlobalKey<_SummaryState> SummaryKey = GlobalKey<_SummaryState>();
+
+class _SummaryState extends State<Summary> {
+  final _lockInit = Lock();
+
+  int total_procured = 0;
+  int total_distributed = 0;
+  int procured_today = 0;
+  int distributed_today = 0;
+
+  Future<void> _futureInit() async {
+    await _lockInit.synchronized(() async {
+      total_procured = await FB().readLadduStockTotal();
+      total_distributed = await FB().readLadduDistSum();
+
+      // get today's stock
+      List<LadduStock> stocks =
+          await FB().readLadduStocks(date: DateTime.now());
+      procured_today = 0;
+      for (var stock in stocks) {
+        procured_today += stock.count;
+      }
+
+      // get today's distribution
+      List<LadduDist> dists = await FB().readLadduDists(date: DateTime.now());
+      distributed_today = 0;
+      for (var dist in dists) {
+        distributed_today += dist.count;
+      }
+    });
+  }
+
+  Future<void> refresh() async {
+    await _futureInit();
+    setState(() {});
+  }
+
+  void restock(int procured) {
+    total_procured += procured;
+    procured_today += procured;
+    setState(() {});
+  }
+
+  Widget _getPieChart() {
+    return PieChart(
+      PieChartData(
+        sections: [
+          PieChartSectionData(
+            color: Colors.blue,
+            value: 40,
+            title: '40%',
+            radius: 50,
+            titleStyle: TextStyle(
+                fontSize: 16, fontWeight: FontWeight.bold, color: Colors.white),
+          ),
+          PieChartSectionData(
+            color: Colors.red,
+            value: 30,
+            title: '30%',
+            radius: 50,
+            titleStyle: TextStyle(
+                fontSize: 16, fontWeight: FontWeight.bold, color: Colors.white),
+          ),
+          PieChartSectionData(
+            color: Colors.green,
+            value: 20,
+            title: '20%',
+            radius: 50,
+            titleStyle: TextStyle(
+                fontSize: 16, fontWeight: FontWeight.bold, color: Colors.white),
+          ),
+          PieChartSectionData(
+            color: Colors.yellow,
+            value: 10,
+            title: '10%',
+            radius: 50,
+            titleStyle: TextStyle(
+                fontSize: 16, fontWeight: FontWeight.bold, color: Colors.white),
+          ),
+        ],
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<void>(
+      future: _futureInit(),
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return const Center(
+            child: CircularProgressIndicator(),
+          );
+        } else if (snapshot.hasError) {
+          return Text('Error: ${snapshot.error}');
+        } else {
+          // return Column(
+          //   crossAxisAlignment: CrossAxisAlignment.start,
+          //   children: [
+          //     Padding(
+          //       padding: const EdgeInsets.only(left: 16.0),
+          //       child: Align(
+          //         alignment: Alignment.centerLeft,
+          //         child: Text("Total laddu packs procured = $total_procured"),
+          //       ),
+          //     ),
+          //     Padding(
+          //       padding: const EdgeInsets.only(left: 16.0),
+          //       child: Align(
+          //         alignment: Alignment.centerLeft,
+          //         child: Text("Total laddu packs served = $total_distributed"),
+          //       ),
+          //     ),
+
+          //   ],
+          // );
+          return _getPieChart();
+        }
+      },
+    );
+  }
+}
